Fix profile services
