+++ src/components/index.ts
@@ -3,6 +3,7 @@
  */
 
 export * from './ToastProvider';
+export * from './account/ProtectedLayoutContent/ProtectedLayoutContent';
 export * from './account/accountPage/Account';
 export * from './account/accountPage/AccountContent';
 export * from './account/accountPage/AccountContentPanel';
@@ -34,7 +35,6 @@ export * from './account/chatPage/ChatModal/ChatModal';
 export * from './account/navigation/BottomNavigation';
 export * from './account/navigation/PageNavigation';
 export * from './account/navigation/SidebarNavigation';
-export * from './account/ProtectedLayoutContent/ProtectedLayoutContent';
 export * from './account/settingsPage/AddCardModal';
 export * from './account/settingsPage/CardPreview';
 export * from './account/settingsPage/DatePicker';
@@ -54,8 +54,8 @@ export * from './layout/header/Header';
 export * from './layout/header/LanguageSwitcher';
 export * from './layout/header/ListDropdown';
 export * from './layout/header/NavDropdown';
-export * from './layout/header/Navigation';
 export * from './layout/header/NavItemRenderer';
+export * from './layout/header/Navigation';
 export * from './layout/header/SettingList';
 export * from './layout/header/SpecialToggle';
 export * from './layout/header/UserAvatar';
@@ -79,9 +79,9 @@ export * from './main/auth/ForgotPassword';
 export * from './main/auth/LoginPageContent';
 export * from './main/auth/RegisterLoginSocial';
 export * from './main/auth/RegisterPageContent';
-export * from './main/auth/safeNextPath';
 export * from './main/auth/Verification';
 export * from './main/auth/VerififyViaEmail';
+export * from './main/auth/safeNextPath';
 export * from './main/donate/Donate';
 export * from './main/donate/DonateCard';
 export * from './main/faq/ContactForm';
