+++ src/services/cloudinary.ts
@@ -1,0 +1,15 @@
+import { fetchFromApi, FetchResult } from '../lib/apiFetcher';
+
+export interface CloudinaryDeleteResponse {
+  result: string;
+  [key: string]: any;
+}
+export const deleteFromCloudinary = async (
+  url: string
+): Promise<FetchResult<CloudinaryDeleteResponse>> => {
+  return await fetchFromApi('/api/cloudinary/delete', {
+    method: 'POST',
+    data: { url },
+    auth: true,
+  });
+};
